Add modularity fo rthe xls-converter
Ref: https://github.com/hl7au/au-fhir-test-data-utils/issues/4
